Dynamic Plugin System

+++ src/berry_mill/mill.py
@@ -41,6 +41,9 @@ class ImageMill:
         """
         Constructor
         """
+        self._bac_appliance_abspth:str = ""
+        self._tmp_backup_dir:str = ""
+
         # Display just help if run alone
         if len(sys.argv) == 1:
             sys.argv.append("--help")
@@ -62,11 +65,15 @@ class ImageMill:
 
         # plugin loader
         plugin.plugins_loader(sub_p)
-
         self.args: argparse.Namespace = p.parse_args()
+        plugin.plugins_args(self.args)
+
         self.cfg: ConfigHandler = ConfigHandler()
         if self.args.config:
             self.cfg.add_config(self.args.config)
+
+        if os.path.exists("project.conf"):
+            self.cfg.add_config("project.conf")
 
     def _add_default_args(self, p: argparse.ArgumentParser) -> None:
         """
@@ -77,7 +84,7 @@ class ImageMill:
         p.add_argument("-d", "--debug", action="store_true", help="turns on verbose debugging mode")
         p.add_argument("-a", "--arch", help="specify target arch")
         p.add_argument("-c", "--config", type=str, help="specify configuration other than default")
-        p.add_argument("-i", "--image", required=True, help="path to the image appliance, if it's not in the current directory")
+        p.add_argument("-i", "--image", help="path to the image appliance, if it's not in the current directory")
         p.add_argument("-p", "--profile", help="select profile for images that makes use of it")
         p.add_argument("--clean", action="store_true", help="cleanup previous build results prior build.")
 
@@ -164,6 +171,28 @@ class ImageMill:
         with open(self._appliance_abspath, "w") as ma:
             ma.write(final_rendered_xml_string)
 
+    def _check_opts(self) -> bool:
+        """
+        Check if options are correct.
+        They are not defined required or not by default
+        """
+        if self.args.subparser_name in ["prepare", "build"]:
+            if self.args.image is None:
+                log.error("Image (--image) is not specified")
+                return False
+        return True
+
+    def _set_appliance_paths(self):
+        """
+        Set appliance paths
+        """
+        self._appliance_path, self._appliance_descr = self._get_appliance_path_info(self.args.image)
+        os.chdir(self._appliance_path)
+
+        self._tmp_backup_dir: str = mkdtemp(prefix="berrymill-tmp-", dir="/tmp")
+        self._appliance_abspath: str = os.path.join(os.getcwd(), self._appliance_descr)
+        self._bac_appliance_abspth: str = os.path.join(self._tmp_backup_dir, self._appliance_descr)
+
     def run(self) -> None:
         """
         Build an image
@@ -173,17 +202,13 @@ class ImageMill:
             print(yaml.dump(self.cfg.config))
             return
 
-        # Set appliance paths
-        self._appliance_path, self._appliance_descr = self._get_appliance_path_info(self.args.image)
-
-        os.chdir(self._appliance_path)
-
-        self._tmp_backup_dir: str = mkdtemp(prefix="berrymill-tmp-", dir="/tmp")
-        self._appliance_abspath: str = os.path.join(os.getcwd(), self._appliance_descr)
-        self._bac_appliance_abspth: str = os.path.join(self._tmp_backup_dir, self._appliance_descr)
-
         kiwip: KiwiParent | None = None
+
+        if not self._check_opts():
+            raise SystemExit()
+
         if self.args.subparser_name == "build":
+            self._set_appliance_paths()
             self._construct_final_appliance()
             # parameter "cross" implies a amd64 host and an arm64 target-arch
             if self.args.cross:
@@ -208,6 +233,7 @@ class ImageMill:
                 no_accel=self.args.no_accel
                 )
         elif self.args.subparser_name == "prepare":
+            self._set_appliance_paths()
             self._construct_final_appliance()
             kiwip = KiwiPreparer(
                 self._appliance_descr,
@@ -218,7 +244,7 @@ class ImageMill:
                 )
         elif self.args.subparser_name is not None:
             log.debug("Calling plugin {}".format(self.args.subparser_name))
-            plugin.registry.call(self.args.subparser_name)
+            plugin.registry.call(self.cfg, self.args.subparser_name)
             return
         else:
             raise argparse.ArgumentError(argument=None, message="No Action defined (build, prepare) or any of available plugins")
@@ -237,6 +263,7 @@ class ImageMill:
         """
         Cleanup Temporary directories and files
         """
-        if (os.path.exists(self._bac_appliance_abspth)):
+        if self._bac_appliance_abspth and (os.path.exists(self._bac_appliance_abspth)):
             shutil.move(self._bac_appliance_abspth, self._appliance_abspath)
-        shutil.rmtree(self._tmp_backup_dir, ignore_errors=True)
+        if self._tmp_backup_dir:
+            shutil.rmtree(self._tmp_backup_dir, ignore_errors=True)
